Fix: src 폴더 삭제 및 Board 오류 해결
