+++ frontend/formyegg/App.js
@@ -4,10 +4,8 @@ import {Provider} from 'react-redux';
 import {PersistGate} from 'redux-persist/integration/react';
 import {persistStore} from 'redux-persist';
 import store from './store';
-import { Interceptor } from './src/common/commonHttp';
+import { Interceptor } from './app/utils/commonHttp';
 import LoginStack from './app/screens/Navigation/LoginStack';
-import BoardScreen from './src/board/Board';
-import RootRouter from './router';
 
 import {
   SafeAreaView,
@@ -27,7 +25,7 @@ function App() {
       <Provider store={store}>
         <PersistGate loading={null} persistor={persistor}>
           <Interceptor>
-            <RootRouter />
+            <LoginStack />
           </Interceptor>
         </PersistGate>
       </Provider>
