--- frontend/formyegg/_App.js
@@ -1,43 +1,0 @@
-import React from 'react';
-import {NavigationContainer} from '@react-navigation/native';
-import {Provider} from 'react-redux';
-import {PersistGate} from 'redux-persist/integration/react';
-import {persistStore} from 'redux-persist';
-import store from './store';
-import RootRouter from './router';
-
-import {
-  SafeAreaView,
-  ScrollView,
-  StatusBar,
-  StyleSheet,
-  Text,
-  View,
-} from 'react-native';
-import BottomTab from './src/common/BottomBar';
-
-export let persistor = persistStore(store);
-
-function App() {
-  return (
-    <NavigationContainer>
-      <Provider store={store}>
-        <PersistGate loading={null} persistor={persistor}>
-          <RootRouter />
-          {/* <BottomTab /> */}
-        </PersistGate>
-      </Provider>
-    </NavigationContainer>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
-
-export default App;
