+++ frontend/formyegg/app/screens/Navigation/LoginStack.js
@@ -6,7 +6,7 @@ import {useSelector} from 'react-redux';
 function LoginStack() {
   const {token} = useSelector(state => state.member);
 
-  return <>{token ? <LoginScreen /> : <TabBottom />}</>;
+  return <>{!token ? <LoginScreen /> : <TabBottom />}</>;
 }
 
 export default LoginStack;
