--- frontend/formyegg/src/common/NavigationService.js
@@ -1,29 +1,0 @@
-import { NavigationActions } from 'react-navigation';
-
-let _navigator;
-
-function setTopLevelNavigator(navigatorRef) {
-    _navigator = navigatorRef;
-}
-
-function navigate(routeName, params) {
-    _navigator.dispatch(
-        NavigationActions.navigate({
-            routeName,
-            params,
-        })
-    )
-    
-}
-
-function back() { 
-    _navigator.dispatch(
-        NavigationActions.back()
-    );
-}
-
-export {
-    navigate,
-    setTopLevelNavigator,
-    back,
-};
