--- frontend/formyegg/src/common/index.js
@@ -1,3 +1,0 @@
-import _NavigationService from './NavigationService';
-
-export const NavigationService = _NavigationService;
